fix: Separate algolia indexing to an independent script in package.json
